backend/requirements.txt
CacheControl==0.12.5
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
click==6.7
firebase-admin==2.13.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
google-api-core==1.5.2
google-auth==1.6.0
google-cloud-core==0.28.1
google-cloud-firestore==0.30.0
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
msgpack==0.5.6
protobuf==3.6.1
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytz==2018.7
requests==2.20.1
rsa==4.0
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.24.1
Werkzeug==0.14.1
python-dotenv==0.10.1
pyyaml
